\name{DaysSince}
\alias{DaysSince}
\title{DaysSince}
\usage{DaysSince(x)}
\description{This function calculates the number of days that have passed since the beginning of the current year.}
\examples{days_passed <- days_since_start_of_year()
cat("Number of days since the beginning of the year:", days_passed, "\n")}
