\name{DaysTilDay}
\alias{DaysTilDay}
\title{DaysTilDay}
\usage{DaysTilDay(x)}
\description{This function calculates the number of days until a specified future date from the current date.}
\examples{days_until <- days_until_date("2023-12-31")
cat("Number of days until December 31, 2023:", days_until, "days\n")}
